add nav visualizer, make visualizer shared util
