+++ src/app/utils/track/SegmentTrackingProvider.ts
@@ -1,10 +1,8 @@
-import Clipboard from "@react-native-community/clipboard"
 import { Analytics } from "@segment/analytics-react-native"
 import { addBreadcrumb } from "@sentry/react-native"
-import { Toast } from "app/Components/Toast/Toast"
-import { unsafe_getDevToggle } from "app/store/GlobalStore"
 import { Platform } from "react-native"
 import Config from "react-native-config"
+import { visualize } from "../visualizer"
 import { isCohesionScreen, TrackingProvider } from "./providers"
 
 export const SEGMENT_TRACKING_PROVIDER = "SEGMENT_TRACKING_PROVIDER"
@@ -48,7 +46,7 @@ export const SegmentTrackingProvider: TrackingProvider = {
     // Events bubbled up from ios native
     if ("screen_name" in info) {
       const { screen_name, ...rest } = info
-      visualize("Screen", screen_name, info)
+      visualize("Screen", screen_name, info, "DTShowAnalyticsVisualiser")
       analytics.screen(screen_name, rest as any)
       return
     }
@@ -57,10 +55,10 @@ export const SegmentTrackingProvider: TrackingProvider = {
       const { action } = info
       if (isCohesionScreen(info)) {
         const { context_screen_owner_type } = info
-        visualize("Screen", context_screen_owner_type, info)
+        visualize("Screen", context_screen_owner_type, info, "DTShowAnalyticsVisualiser")
         analytics.screen(context_screen_owner_type, info as any)
       } else {
-        visualize("Track", action, info)
+        visualize("Track", action, info, "DTShowAnalyticsVisualiser")
         analytics.track(action, info as any)
       }
       return
@@ -68,21 +66,21 @@ export const SegmentTrackingProvider: TrackingProvider = {
 
     if ("action_type" in info) {
       const { action_type, ...rest } = info
-      visualize("Track", action_type, info)
+      visualize("Track", action_type, info, "DTShowAnalyticsVisualiser")
       analytics.track(action_type, rest as any)
       return
     }
 
     if ("name" in info) {
       const { name, ...rest } = info
-      visualize("Track", name, info)
+      visualize("Track", name, info, "DTShowAnalyticsVisualiser")
       analytics.track(name, rest as any)
       return
     }
 
     if ("context_screen" in info) {
       const { context_screen, ...rest } = info
-      visualize("Screen", context_screen, info)
+      visualize("Screen", context_screen, info, "DTShowAnalyticsVisualiser")
       analytics.screen(context_screen, rest as any)
       return
     }
@@ -90,7 +88,7 @@ export const SegmentTrackingProvider: TrackingProvider = {
     // default check events from ios native
     if ("event_name" in info) {
       const { event_name, ...rest } = info
-      visualize("Track", event_name, info)
+      visualize("Track", event_name, info, "DTShowAnalyticsVisualiser")
       analytics.track(event_name, rest as any)
       return
     }
@@ -98,31 +96,4 @@ export const SegmentTrackingProvider: TrackingProvider = {
     console.warn("oh wow, we are not tracking this event!! we should!", { info })
     assertNever(info)
   },
-}
-
-const visualize = (type: string, name: string, info: { [key: string]: any }) => {
-  if (!unsafe_getDevToggle("DTShowAnalyticsVisualiser")) {
-    return
-  }
-
-  const title = `${type}: ${name}`
-  const message = JSON.stringify(info, null, 2)
-
-  Toast.show(title, "top", {
-    onPress: ({ showActionSheetWithOptions }) => {
-      showActionSheetWithOptions(
-        {
-          title,
-          message,
-          options: ["Copy description", "Continue"],
-          cancelButtonIndex: 1,
-        },
-        (buttonIndex) => {
-          if (buttonIndex === 0) {
-            Clipboard.setString(message)
-          }
-        }
-      )
-    },
-  })
 }
